archive/containers/styletts2/requirements.txt
SoundFile
munch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
phonemizer
einops
einops-exts
tqdm
typing-extensions
git+https://github.com/resemble-ai/monotonic_align.git